Move is_top from pyk.kastManip to pyk.prelude.ml
